order is integer
git-svn-id: svn+ssh://localhost/Home/stat/WorkSpace/dti-svn/trunk@278 8193cdef-c62e-4131-a968-7dcc8d47c1f7

+++ R/tracking.r
@@ -255,7 +255,7 @@ function(obj, roix=NULL, roiy=NULL, roiz=NULL, method="LINEPROP", minanindex=0.3
   
   dd <- .Call("interface_tracking_mixtensor",
               as.double(ex$andir), # dim = c(3, maxorder, dimx, dimy, dimz)
-              as.double(ex$order), # NEW! dim = c(dimx, dimy, dimz)
+              as.integer(ex$order), # NEW! dim = c(dimx, dimy, dimz)
               as.double(ex$gfa),    # dim = c(dimx, dimy, dimz)
               as.double(ex$mix),    # NEW! dim = c(maxorder, dimx, dimy, dimz)
               as.integer(maxorder), # NEW!
